requirements.txt
h5py==3.12.1
joblib==1.4.2
keras==3.8.0
matplotlib==3.9.0
numpy==2.0.2
pandas==2.2.2
Platypus-Opt==1.2.0
scikeras==0.13.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
skrebate==0.62
tensorflow==2.18.0